Extract HTTP behavior into object
